minor nits and rename
